Update Lua client and Python server for enhanced logging, protocol improvements, and file handling
- Changed temporary file format in `nitrogen_client.lua` from BMP to PNG.
- Added file size logging and JSON header length inclusion in Lua client.
- Improved error handling and logging in `serve.py` with `flush=True` and traceback output.
- Ensured proper copying of reshaped arrays for image processing in Python server.
- Introduced slight delay for thread initialization in the TCP server setup.

+++ nitrogen/flow_matching_transformer/nitrogen.py
@@ -435,7 +435,7 @@ class NitroGen(torch.nn.Module):
 
         B, T = sa_token_ids.shape
         sa_embs = torch.full(
-            size=(B, T, self.hidden_size), fill_value=0.0, dtype=vision.dtype, device=vision.device
+            size=(B, T, self.hidden_size), fill_value=0.0, dtype=action.dtype, device=vision.device
         )
 
         # Project state.
